Moving all paparazzi tests into individual module. (#146)
