+++ .github/workflows/paparazzi_tests.yml
@@ -19,4 +19,4 @@ jobs:
         uses: ./.github/actions/workflow_setup
 
       - name: Run Tests
-        run: ./gradlew verifyPaparazziDebug
+        run: ./gradlew :test:paparazzi:verifyPaparazziDebug
