+++ .github/workflows/unit_tests.yml
@@ -19,4 +19,4 @@ jobs:
         uses: ./.github/actions/workflow_setup
 
       - name: Run Tests
-        run: ./gradlew test
+        run: ./gradlew test -x :test:paparazzi:test
