+++ .gitignore
@@ -9,3 +9,4 @@
 .cxx
 local.properties
 .kotlin/
+kover.artifact
